Revert "Merge branch 'issue48' (and issue 81) into dev"
This reverts commit 5204218c5ee99d3577fee8ceff17eb7b99ce33a6, reversing
changes made to b11082b3b1af69ef3130166fcc2b30b2fe49a8bf.
