+++ ast.go
@@ -226,34 +226,6 @@ func (n *lsAttrNode) execute() (interface{}, error) {
 	return nil, nil
 }
 
-type lsAttrGenericNode struct {
-	path     node
-	argFlags map[string]interface{}
-}
-
-func (n *lsAttrGenericNode) execute() (interface{}, error) {
-	arg := ""
-	path, err := AssertString(&n.path, "Path")
-	if err != nil {
-		return nil, err
-	}
-
-	if len(n.argFlags) > 1 {
-		return nil,
-			fmt.Errorf("This command accepts a single '-s' argument only")
-	}
-
-	if len(n.argFlags) > 0 && n.argFlags["s"] == nil {
-		return nil,
-			fmt.Errorf("This command accepts a single '-s' argument only")
-	} else {
-		arg = n.argFlags["s"].(string)
-	}
-
-	cmd.LSATTR(path, arg)
-	return nil, nil
-}
-
 type getUNode struct {
 	path node
 	u    node
@@ -693,9 +665,9 @@ func (n *easyUpdateNode) execute() (interface{}, error) {
 }
 
 type lsObjNode struct {
-	path     node
-	entity   int
-	argFlags map[string]interface{}
+	path   node
+	entity int
+	flag   node
 }
 
 func (n *lsObjNode) execute() (interface{}, error) {
@@ -707,107 +679,17 @@ func (n *lsObjNode) execute() (interface{}, error) {
 	if !ok {
 		return nil, fmt.Errorf("Path should be a string")
 	}
+	flag, err := AssertString(&n.flag, "Flag")
+	if err != nil {
+		return nil, err
+	}
 
-	args := n.argFlags
-	switch len(args) {
-	case 0:
-		return cmd.LSOBJECT(path, n.entity, false), nil
-	case 1:
-		//check for -r or -s
-		if _, ok := args["r"]; ok {
-			if args["r"] != nil {
-				return nil, fmt.Errorf("-r takes no arguments")
-			}
-			return cmd.LSOBJECTRecursive(path, n.entity, false), nil
-
-		} else if _, ok := args["s"]; ok {
-			if IsStringArr(args["s"]) {
-				msg := "Too many arguments supplied, -s only takes one"
-				return nil, fmt.Errorf(msg)
-			}
-			if !IsString(args["s"]) {
-				msg := "Please provide a string argument for '-s'"
-				return nil, fmt.Errorf(msg)
-			}
-
-			objs := cmd.LSOBJECT(path, n.entity, true)
-			sorted := cmd.SortObjects(&objs, args["s"].(string))
-			sorted.Print()
-			return objs, nil
-
-		} else if _, ok := args["f"]; ok {
-			var arr []string
-			if IsString(args["f"]) {
-				arr = []string{args["f"].(string)}
-			} else if IsStringArr(args["f"]) {
-				arr = args["f"].([]string)
-			} else {
-				msg := "Please provide a string(s) argument for '-f'"
-				return nil, fmt.Errorf(msg)
-			}
-
-			objs := cmd.LSOBJECT(path, n.entity, true)
-			cmd.DispWithAttrs(&objs, &arr)
-			return nil, nil
-
-		} else {
-			msg := "Unknown argument received. You can only use '-r' or '-s'"
-			return nil, fmt.Errorf(msg)
-		}
-	case 2:
-		//check for -r and (-s  or -f)
-		for i := range args {
-			if !IsAmongValues(i, &[]string{"r", "s", "f"}) {
-				msg := "Unknown argument received." +
-					" You can only use '-r' or '-s' or '-f'"
-				return nil, fmt.Errorf(msg)
-			}
-		}
-
-		//Error check can't have -s and -f together
-		if _, ok := args["s"]; ok {
-			if _, ok := args["f"]; ok {
-				msg := "Invalid argument order received." +
-					" You can only use '-r' in combination with '-s' or '-f'"
-				return nil, fmt.Errorf(msg)
-			}
-		}
-
-		if args["r"] != nil {
-			return nil, fmt.Errorf("-r takes no arguments")
-		}
-
-		objs := cmd.LSOBJECTRecursive(path, n.entity, true)
-
-		if _, ok := args["s"]; ok {
-			if IsString(args["s"]) {
-				sorted := cmd.SortObjects(&objs, args["s"].(string))
-				sorted.Print()
-				return objs, nil
-			} else {
-				msg := "Please provide a string argument for '-s'"
-				return nil, fmt.Errorf(msg)
-			}
-
-		}
-
-		//Else '-f'
-		var attrs []string
-		if IsString(args["f"]) {
-			attrs = []string{args["f"].(string)}
-		} else if IsStringArr(args["f"]) {
-			attrs = args["f"].([]string)
-		} else {
-			msg := "Please provide a string(s) argument for '-f'"
-			return nil, fmt.Errorf(msg)
-		}
-
-		cmd.DispWithAttrs(&objs, &attrs)
-		return objs, nil
-
-	default:
-		//Return err
-		msg := "Too many arguments. You can only use '-r' or '-s'"
+	if flag == "r" {
+		return cmd.LSOBJECTRecursive(path, n.entity), nil
+	} else if flag == "" {
+		return cmd.LSOBJECT(path, n.entity), nil
+	} else {
+		msg := "Unrecognised flag. You can only use '-r'"
 		return nil, fmt.Errorf(msg)
 	}
 
@@ -906,9 +788,17 @@ type selectChildrenNode struct {
 }
 
 func (n *selectChildrenNode) execute() (interface{}, error) {
-	paths, err := evalNodeArr[string](&n.paths, []string{})
-	if err != nil {
-		return nil, err
+	var paths []string
+	for i := range n.paths {
+		v, err := n.paths[i].execute()
+		if err != nil {
+			return nil, err
+		}
+		path, ok := v.(string)
+		if !ok {
+			return nil, fmt.Errorf("")
+		}
+		paths = append(paths, path)
 	}
 	v, err := cmd.SetClipBoard(paths)
 	if err != nil {
