+++ interpreter/lexer.nex
@@ -95,6 +95,8 @@
 /lsac/ {  printToks("TOK_LSAC"); return TOK_LSAC}
 /lspanel/ {  printToks("TOK_LSPANEL"); return TOK_LSPANEL}
 /lscorridor/ {  printToks("TOK_LSCORRIDOR"); return TOK_LSCORRIDOR}
+/lsu/ {  printToks("TOK_LSU"); return TOK_LSU}
+/lsslot/ {  printToks("TOK_LSSLOT"); return TOK_LSSLOT}
 /lsenterprise/ {  printToks("TOK_LSENTERPRISE"); return TOK_LSENTERPRISE}
 /tree/ { printToks("TOK_TREE"); return TOK_TREE}
 /lsog/ { printToks("TOK_LSOG"); return TOK_LSOG}
