+++ interpreter/parser.y
@@ -14,10 +14,14 @@ var _ = l.GetInfoLogger() //Suppresses annoying Dockerfile build error
   n int
   s string
   f float64
+  sarr []string
   ast *ast
   node node
+  boolNode boolNode
+  numNode numNode
   nodeArr []node
-  stringArr []string
+  arr []interface{}
+  mapArr []map[int]interface{}
   mapVoid map[string]interface{}
 }
 
@@ -37,7 +41,7 @@ var _ = l.GetInfoLogger() //Suppresses annoying Dockerfile build error
        TOK_CLR TOK_GREP TOK_LS TOK_TREE
        TOK_LSOG TOK_LSTEN TOK_LSSITE TOK_LSBLDG
        TOK_LSCAB TOK_LSSENSOR TOK_LSAC TOK_LSPANEL
-       TOK_LSCORRIDOR TOK_GETU
+       TOK_LSCORRIDOR TOK_LSU TOK_LSSLOT TOK_GETU
        TOK_LSROOM TOK_LSRACK TOK_LSDEV TOK_LSENTERPRISE
        TOK_ATTRSPEC TOK_GETSLOT
        TOK_COL TOK_SELECT TOK_LBRAC TOK_RBRAC
@@ -59,9 +63,7 @@ var _ = l.GetInfoLogger() //Suppresses annoying Dockerfile build error
        
 %type <n> LSOBJ_COMMAND
 %type <s> OBJ_TYPE COMMAND UI_TOGGLE
-%type <nodeArr> WNARG GETOBJS WORD_CONCAT
-%type <mapVoid> ARGACC
-%type <stringArr> FARGACC
+%type <nodeArr> WNARG GETOBJS
 %type <node> OCCR PATH PHYSICAL_PATH STRAY_DEV_PATH EXPR CONCAT CONCAT_TERM stmnt st2 IF 
        EXPR_NOQUOTE ARRAY ORIENTATION EXPR_NOQUOTE_NOCOL CONCAT_NOCOL CONCAT_TERM_NOCOL EXPR_NOQUOTE_COMMON
 //%type <mapVoid> EQUAL_LIST
@@ -87,18 +89,17 @@ st2:    {$$=nil}
 
 stmnt:   TOK_GET PATH {$$=&getObjectNode{$2}}
        //| TOK_GET OBJ_TYPE EQUAL_LIST {$$=&searchObjectsNode{$2, $3}}
-       
-       //NORMAL LSOBJ COMMANDS
-       | LSOBJ_COMMAND PATH {$$=&lsObjNode{$2, $1,nil}}
-       | LSOBJ_COMMAND {$$=&lsObjNode{&pathNode{&strLeaf{"."}, STD}, $1, nil}}
-       
-       //ARGUMENT LSOBJ COMMANDS 
-       | LSOBJ_COMMAND PATH ARGACC {$$=&lsObjNode{$2, $1, $3}}
-       | LSOBJ_COMMAND ARGACC {$$=&lsObjNode{&pathNode{&strLeaf{"."}, STD}, $1, $2}}
 
-       //ARGUMENT TYPE LS COMMANDS
-       | TOK_LS ARGACC {$$=&lsAttrGenericNode{&pathNode{&strLeaf{"."}, STD}, $2}}
-       | TOK_LS PATH ARGACC {$$=&lsAttrGenericNode{$2, $3}}
+       | LSOBJ_COMMAND PATH {$$=&lsObjNode{$2, $1, &strLeaf{""}}}
+       | LSOBJ_COMMAND {$$=&lsObjNode{&pathNode{&strLeaf{"."}, STD}, $1, &strLeaf{""}}}
+       | LSOBJ_COMMAND PATH TOK_MINUS TOK_WORD {$$=&lsObjNode{$2, $1, &strLeaf{$4}}}
+       | LSOBJ_COMMAND TOK_MINUS TOK_WORD {$$=&lsObjNode{&pathNode{&strLeaf{"."}, STD}, $1, &strLeaf{$3}}}
+       | LSOBJ_COMMAND TOK_MINUS TOK_WORD PATH  {$$=&lsObjNode{$4, $1, &strLeaf{$3}}}
+
+       | TOK_LSU PATH {$$=&lsAttrNode{$2, "heightu"}}
+       | TOK_LSU {$$=&lsAttrNode{&pathNode{&strLeaf{"."}, STD}, "heightu"}}
+       | TOK_LSSLOT PATH {$$=&lsAttrNode{$2, "slot"}}
+       | TOK_LSSLOT {$$=&lsAttrNode{&pathNode{&strLeaf{"."}, STD}, "slot"}}
 
        | TOK_GETU {x:=&pathNode{&strLeaf{"."}, STD}; y:=&intLeaf{0};$$=&getUNode{x, y}}
        | TOK_GETU PATH {$$=&getUNode{$2, &intLeaf{0}}}
@@ -294,30 +295,6 @@ WNARG: EXPR TOK_COMMA WNARG {x:=[]node{$1}; $$=append(x, $3...)}
        |EXPR  {x:=[]node{$1}; $$=x}
 ; 
 
-WORD_CONCAT: TOK_WORD TOK_COMMA WORD_CONCAT {$$=append([]node{&strLeaf{$1}}, $3...)}
-           | {$$=nil} 
-           | TOK_WORD {$$=[]node{&strLeaf{$1}}}
-;
-//Argument Accumulator
-ARGACC: TOK_MINUS TOK_WORD {$$=map[string]interface{}{$2:nil}}
-       |TOK_MINUS TOK_WORD ARGACC {$3[$2]=nil;$$=$3}
-       |TOK_MINUS TOK_WORD TOK_WORD ARGACC {$4[$2]=$3;$$=$4}
-       |TOK_MINUS TOK_WORD TOK_WORD FARGACC {
-              if $4 == nil {
-                     $$=map[string]interface{}{$2:$3}
-              } else {
-                     $$=map[string]interface{}{$2: append([]string{$3},$4...)}
-              }
-       }
-
-;
-
-//Argument Accumulator for '-f'
-FARGACC: TOK_WORD FARGACC {$$=append([]string{$1}, $2...)}
-        | {$$=nil}
-;
-
-
 GETOBJS: PATH TOK_COMMA GETOBJS {x:=[]node{$1}; $$=append(x, $3...)}
        | PATH {x:=[]node{$1}; $$=x}
 ;
@@ -343,8 +320,8 @@ COMMAND: TOK_LINK{$$="link"} | TOK_UNLINK{$$="unlink"} | TOK_CLR{$$="clear"} | T
        | TOK_LSTEN{$$="lsten"} | TOK_LSSITE{$$="lssite"} | TOK_LSBLDG{$$="lsbldg"} | TOK_LSROOM{$$="lsroom"} 
        | TOK_LSRACK{$$="lsrack"} | TOK_LSDEV{$$="lsdev"} | TOK_MINUS{$$="-"} | TOK_TEMPLATE{$$=".template"}
        | TOK_CMDS{$$=".cmds"} | TOK_VAR{$$=".var"} | TOK_PLUS{$$="+"} | TOK_EQUAL{$$="="} 
-       | TOK_GREATER{$$=">"} | TOK_DRAWABLE{$$="drawable"}
-       | TOK_GETU{$$="getu"} | TOK_GETSLOT{$$="getslot"}
+       | TOK_GREATER{$$=">"} | TOK_DRAWABLE{$$="drawable"} | TOK_LSU{$$="lsu"} 
+       | TOK_LSSLOT{$$="lsslot"} | TOK_GETU{$$="getu"} | TOK_GETSLOT{$$="getslot"}
        | TOK_GREP {$$="grep"}
 ;
